[project @ 2002-07-30 22:51:36 by craigmcc] [4.8][4.9][4.10][4.11][SelectItem][UISelectItem][UISelectItems] [UISelectBase][UISelectMany][UISelectOne]
Introduce new UISelectItem and UISelectItems that represent a single
instance of SelectItem, or one or more instances, respectively.  Zero or
more instances of these components may be nested inside a UISelectMany
or UISelectOne component to configure the set of available options.

Remove obsolete "items" and "itemsModelReference" properties on
UISelectMany and UISelectOne.

Original author: craigmcc
Date: 2002-07-30 22:51:37+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@563 761efcf2-d34d-6b61-9145-99dcacc3edf1