[project @ 2002-08-15 16:29:27 by craigmcc] [UICommand]
Update encodeBegin()/encodeEnd() to render an HTML <button> element,
as per the default behavior of the proposed standard RenderKit.

Update decode() behavior to look up the command name and the form name
(for the FormEvent to be enqueued) via currentValue() on the corresponding
UICommand and UIForm components, so that the names can be determined
dynamically via the model if desired.

Original author: craigmcc
Date: 2002-08-15 16:29:27+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@670 761efcf2-d34d-6b61-9145-99dcacc3edf1