[project @ 2003-05-09 22:58:52 by craigmcc] Make the "Faces_Basic.jsp" page work under Tomcat 5.0.2.
The reason for this is a little obscure, but we need to understand it.
In JSP 1.2, there was a very obscure requirement (not actually implemented
by Tomcat 4.0.x or 4.1.x) that custom tags using an "id" attribute would
implicitly create a scripting variable in the generated page with a
corresponding name.  This wasn't detected by the TCK, but some containers
actually did implement this requirement correctly.

In JSP 2.0, the requirement was clarified, and I imagine that the TCK
will now catch violations.  However, it has a side effect -- you need to
use a value for the "id" attribute that is a valid Java variable name,
or else you'll get a compile error.  We were using "double" ...

Original author: craigmcc
Date: 2003-05-09 22:58:52+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@1403 761efcf2-d34d-6b61-9145-99dcacc3edf1