[project @ 2003-09-19 00:57:06 by craigmcc] [JSFA115] Flesh out ValueHolderSupport, including its support for saving and restoring its own state as an attached object.
[JSFA115] Make all components that implement ValueHolder delegate this
functionality to an instance of ValueHolderSupport, and make sure this
instance variable is saved and restored as an attached object.

[JSFA115] Add a new test case in ValueHolderTestCaseBase that ensures
all of the properties reflected in the ValueHolder interface (including
a precisely configured Converter instance) are correctly saved and
restored.  This test is inherited by all of the component test cases
for components that implement ValueHolder, so saving/restoring the
attached object will be tested in all of those cases.

RI IMPACT:  None.

Original author: craigmcc
Date: 2003-09-19 00:57:17+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@1914 761efcf2-d34d-6b61-9145-99dcacc3edf1