[project @ 2003-09-24 23:58:53 by craigmcc] Add initial version of a test case for dynamically adding components to the component tree at runtime.  Our current restriction on this use case is that you can only do so underneath a component that returns true for the rendersChildren property, so that's what we're testing here.
Writing this test case uncovered a potential issue with the way we've tried to
deal with the Reload button -- it seems that doing two GET transactions on
two different pages (/faces/jsp/jsp-dynamic-00.jsp and
/faces/jsp/jsp-dynamic-01.jsp) with no query parameters triggers the reload
detection logic and doesn't switch trees.  As a workaround for now, the
individual tests that this messes up are commented out under a PENDING
comment.  Justyna is investigating the reload detection algorithm.

For me, with JSP-API head, this code passes all systests, but there are some
cactus based tests that still fail.

Original author: craigmcc
Date: 2003-09-24 23:59:00+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@1955 761efcf2-d34d-6b61-9145-99dcacc3edf1