[project @ 2003-10-10 14:31:53 by eburns] [JSFA118]
Spec that any errors during startup or parsing of the config file must
cause the FacesServlet.init() method to throw ServletException, causing the
servlet to not be placed into service.

RI IMPACT:

Make sure that errors in the config file cause FacesServlet.init() to throw
ServletException.  Have some unit tests that prove this is the case.

Original author: eburns
Date: 2003-10-10 14:31:53+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@2095 761efcf2-d34d-6b61-9145-99dcacc3edf1