[project @ 2003-11-05 01:01:11 by craigmcc] Scarab Issue:  [JSFA29] DataGrid
Make UIData act like DataModel, with respect to setting a rowIndex that is
>= the actual number of available rows.  This no longer causes an exception.
Add UIData.isRowAvailable() to allow callers to determine whether there actually
is such a row.

RI IMPACT:  TableRenderer needs to be modified to not rely on
IndexOutOfBoundsException to tell when we scrolled off the end
of the available rows.  (I will fix this in the next commit.)

Original author: craigmcc
Date: 2003-11-05 01:01:12+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@2308 761efcf2-d34d-6b61-9145-99dcacc3edf1