[project @ 2003-11-06 23:01:41 by eburns] [JSFA80] displaying errors and messages
Spec work for the EG issue described at
<http://wspack.sfbay.sun.com/jsf/eg-archive/3594.html> or
<http://archives.java.sun.com/cgi-bin/wa?A2=ind0311&L=jsr-127-eg&F=&S=&X=091F5169A37F6AE36C&Y=ed.burns@sun.com&P=2151>.

RI IMPACT:

For UIMessages

Implement a new renderer, Messages, that knows how to render a
UIMessages instance.  You can probably cop code from ErrorsRenderer.

Make sure to look at the severity of the message and output the
appropriate value for the *style and *styleClass attributes.  For
example, if the message to be currently rendered has a severity of WARN,
then the markup must reflect the values of the warnStyle and
warnStyleClass attributes, if present.

For UIMessage

This one is pretty much a matter of removing the MessageRenderer,
renaming the ErrorsRenderer to be MessageRenderer, and conforming to the
new UIMessage contract.  The same bit about severity and style as for
UIMessages applies.  The existing logic for finding the component that
currently resides in ErrorsRenderer must be preserved.

Other tasks:

Remove output_errors and output_message, and accompanying artifacts.
This means editing the standard-html-renderkit.xml and
standard-components.xml file, as well as removing or renaming the
appropriate renderers.  I elected not to do these removals as part of
the API checkin to preserve the working state of the tree.  I want the
removal to be done atomically with the checkin of the implementation of
the new feature.

SECTION: API Changes

A jsf-api/src/javax/faces/component/UIMessages.java
A jsf-api/src/javax/faces/component/UIMessage.java

- new components

M jsf-api/doc/standard-components.xml

- Add component-type Messages and Message

- Add messages attributes to Table renderer-type

M jsf-api/doc/standard-html-renderkit.xml

- Add component-type Message and MessagesTable

- Add attributes to renderer-type Message and Table

SECTION: RI Changes

M jsf-ri/src/com/sun/faces/standard-html-renderkit-impl.xml

- new renderer type: Messages.

Original author: eburns
Date: 2003-11-06 23:01:45+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@2324 761efcf2-d34d-6b61-9145-99dcacc3edf1