[project @ 2003-11-07 15:09:47 by eburns] [VBL]
Workaround for bug introduced with advent of UIComponentBase.facesContext ivar.

This ivar is never cleared out in the "save state in server case".

I put in a PENDING() comment to revisit this.

Current workaround is to always use ThreadLocal storage until we have
a way for a UIComponentBase instance to reliably know when the end of the
Request Processing Lifecycle has occurred.

Original author: eburns
Date: 2003-11-07 15:09:47+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@2328 761efcf2-d34d-6b61-9145-99dcacc3edf1