[project @ 2005-03-11 18:56:39 by rogerk] Issue number: 6 Obtained from: Submitted by: rogerk Reviewed by:  edburns
This change bundle was needed because of a backwards compatability issue
discovered by adding the "label" placeholder in standard messages that
are used by custom validators/converters.
"Out of the box", applications that have written a custom validator,
and use an existing mechanism to load the standard message (like MessageFactory)
will display this message as:
       "{2}: Validation Error: Specified attribute is not between the expected values of 1 and 10"
which breaks backwards compatability.

--- javadoc changes to refer to own NOT_IN_RANGE message
    (not the deprecated Validator one)
--- added NOT_IN_RANGE message constant
--- throw own NOT_IN_RANGE (not Validator)
--- javadoc changes to deprecate NOT_IN_RANGE

Original author: rogerk
Date: 2005-03-11 18:56:40+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn@3102 761efcf2-d34d-6b61-9145-99dcacc3edf1