M      jsf-ri/src/com/sun/faces/application/resource/ResourceImpl.java M      jsf-demo/ezcomp/src/main/webapp/resources/ezcomp/css_master.css
Make it so resources that need to refer to other resources within the
same library can use the syntax "#{resource['this:progressBar.css']}" to
do so.  For example:

in CSS file in library ezcomp, this expression

@import url(#{resource['this:progressBar.css']});

Is equivalent to this expression

@import url(#{resource['ezcomp:progressBar.css']});

But the benefit is that you don't need to bake the name of the library
into the code of the library.

Next step is to allow a composite component PDL file to refer to
inner-library resources in the same way.




git-svn-id: https://svn.java.net/svn/mojarra~svn@4825 761efcf2-d34d-6b61-9145-99dcacc3edf1