Fix for issue 814 and more.
Issue 814 reported a failure in TestStateManagerImpl when validating StateManager.getViewState() with 
client-side state saving.

The root cause of the issue there was the fact that StringBuilderWriter, wasn't casting the provided int arg
in write(int) to char.  Test case probably didn't fail before due to some boundry condition.  At any rate,
it's resolved now.

Also, UIViewRoot is no longer a ComponentSystemEventListener.  Instead, override processRestoreState()
and notify any UIViewRoot-local PhaseListeners after the call to super.processRestoreState() has completed.

Doing the above removes the need to register UIViewRoot as a listener in RestoreViewPhase.

In the process of testing, we discovered a recursion issue restoring SystemEventListener state for the case
where a UIComponent is registered to itself as a listener for a particular event.  The changes included with
this commit resolve this issue, so there will be differences between this commit and the change bundle 
included with 814.

git-svn-id: https://svn.java.net/svn/mojarra~svn@5546 761efcf2-d34d-6b61-9145-99dcacc3edf1