Fix for MOJARRA-836
The problem was due to the fact that the value was being passed as a literal.
In JSP this works fine as the JSP compile will create a ValueExpression for the literal and apply the coercion as specified in the TLD.  

In the facelets case, when a literal value is encountered, it creates a LiteralValueMetaData instance that merely passes the value in the view to the local value of the component.  When UISelectBoolean.isSelected() is called, the String value causes a ClassCastException.

The fix removes LiteralValueMetaData and creates an expression no matter if it's a literal or not, but special cases the UISelectBoolean case so that the expected type is Boolean.

git-svn-id: https://svn.java.net/svn/mojarra~svn@6128 761efcf2-d34d-6b61-9145-99dcacc3edf1