Issue: 140-Flash
jsf-ri/conf/share/jstl-core.tld

- TLDDocs for c:set

Sets the result of an expression evaluation based on the value of the
attributes. The handler for this tag must ensure that the action of
evaluating the expression happens during the rendering of the tree, not
during the build of the tree.

This handler operates in one of two ways.

The user has set "var", "value" and (optionally) "scope" attributes.

The user has set "target", "property", and "value" attributes.

The first case takes precedence over the second.

Case 1

The presence or absence of a value for the "scope" attribute affects how
the "var" attribute must be handled. If no "scope" attribute is present,
the implementation must get the current ELContext, get its
VariableMapper, and call setVariable(), passing the value of the "var"
attribute as the first argument and the un-evaluated expression (or
literal) from the "value" attribute as the second.

If "scope" is present, but has a zero length, TagException is thrown
with an informative error message, ensuring page location information is
saved.

If "scope" is present and is a literal, let scopeStr be the value of the
literal. If "scope" is present and is an expression, let scopeStr be the
evaluated value of the expression. If scopeStr is "page", throw
TagException as described above. Otherwise, if scopeStr is "request",
"session", "application", or "view" (without the quotes) append the
String "Scope" (without the quotes) to scopeStr. Let varStr be the
evaluated value of the "var" attribute. Let leftHandSide be
ValueExpression created with the expression string "#{" + scopeStr + "."
+ varStr + "}". Call its setValue() method, passing the evaluated value
of the "value" attribute.

Case 2

In this case "var" must not be set, and "target", "value" and "property"
all must be set.

If "target" is a literal string throw TagException as described
above. If "property" is a literal string, let propertyStr be its literal
value, otherwise, let propertyStr be its evaluated value. Let
targetValue be the result of calling getValue() on the ValueExpression
from the "target" attribute. Obtain the ELContext from the current
FacesContext. Obtain the ELResolver from the ELContext. Call setValue()
on the ELResolver, passing the appropriate ELContext as the first
argument, targetValue as the second argument, propertyStr as the third
argument, and the evaluated result of the "value" attribute as the
fourth argument.



git-svn-id: https://svn.java.net/svn/mojarra~svn@6259 761efcf2-d34d-6b61-9145-99dcacc3edf1