Fix for MOJARRA-998
Enable support of nested #{compositeComponent.attrs.} expressions while be independant 
of nesting levels.

Consider:

Using Page:
=====================================
<ez:testcomponent1 backingBean="#{inout}" testObject="#{inout}"/>


testcomponent1.xhtml
====================================
<composite:implementation>
<t:testcomponent2
nestedTestObject="#{compositeComponent.attrs.testObject}"
nestedBackingBean="#{compositeComponent.attrs.backingBean}" />
</composite:implementation>


testcomponent2.xhtml
===================================
<composite:implementation>

    <div style="border: solid 4px silver;">
        <h2>Level 2</h2>
        <div>Bean: #{compositeComponent.attrs.nestedBackingBean}</div>
        <div>Test Object: #{compositeComponent.attrs.nestedTestObject}</div>
    </div>

</composite:implementation>

=================================================

Before this fix, none of the expression in testcomponen2.xhtml would evaluate as 
the expression #{compositeComponent.attrs.nestedBackingBean} would refer to
the expression #{compositeComponent.attrs.testObject} and the #{compositeComponent
key work always refers to the current composite component.  The component
representing testcomponent2.xhtml has no testObject or backingBean expressions.

To solve this, we add a ComponentSystemEventListener for the PostAddToViewEvent 
to all CompositeComponets.  When this listener is invoked, it will check to see
if it has a composite component parent somewhere in the hierarchy.  If it does,
it will iterate through the expressions of the event source and parse them
out for potential attribute names within the parent.  If an attribute value within
the parent is found and is an Expression, we'll copy that expression from the 
parent to the child.  Doing so means when testcomponent2.xhtml evaluates 
#{compositeComponent.attrs.nestedBackingBean} the expression actually being
evaluated is #{inout}


git-svn-id: https://svn.java.net/svn/mojarra~svn@6703 761efcf2-d34d-6b61-9145-99dcacc3edf1