Issue 1032 - Component Behavior API refactoring
The following classes have been renamed:

- Behavior class -> ClientBehavior interface
- BehaviorHolder->ClientBehaviorHolder
- BehaviorContext->ClientBehaviorContext
- BehaviorRenderer->ClientBehaviorRenderer
- BehaviorBase->ClientBehaviorBase
- BehaviorHint->ClientBehaviorHint

The following methods have also been renamed:

- BehaviorHolder.addBehavior() -> ClientBehaviorHolder.addClientBehavior()
- BehaviorHolder.getBehaviors() -> ClientBehaviorHolder.getClientBehaviors()
- RenderKit.addBehaviorRenderer() -> RenderKit.addClientBehaviorRenderer()
- RenderKit.getBehaviorRendererTypes() -> RenderKit.getClientBehaviorRendererTypes().

And we now have a new base javax.faces.component.behavior.Behavior (empty) interface.  This serves as the root of the behavior API hierarcy.  For the moment, this only type of behaviors that we support our ClientBehaviors, but we expect to add to this in 2.1.

git-svn-id: https://svn.java.net/svn/mojarra~svn@6801 761efcf2-d34d-6b61-9145-99dcacc3edf1