Fix for issue 962.
I was unable to reproduce the issue with
com.sun.faces.disableUnicodeEscaping set to auto with Facelets 1.1.14, but I was able to reproduce the problem with our implmentation.

The root cause was due to some refactoring in the compiler to use the FacesContext attribute map to store the encoding, but the logic that checked the encoding in the ViewHandler was still looking in the request.

Also changed the default value for com.sun.faces.disableUnicodeEscaping to be auto.

git-svn-id: https://svn.java.net/svn/mojarra~svn@7319 761efcf2-d34d-6b61-9145-99dcacc3edf1