Fix for issue 1185.
If a component is dynamically added, check the remove list for an existing matching client ID and remove it from the remove list.  Do the same when removing a component, check to see if it's in the added list and if so, remove it from there.

Update UIComponentBase ChildrenList actions that cause a component to be removed.  Previous version, only ChildrenList.remove(UIComponent) would trigger a PreRemoveFromViewEvent.  Logic has been updated to trigger the event when the parent of the component is set to null instead of updating all remove points from ChildrenList.

git-svn-id: https://svn.java.net/svn/mojarra~svn@7495 761efcf2-d34d-6b61-9145-99dcacc3edf1