Fix for Bug 1184
Previously, selectManyCheckbox did not work with ajax calls.  The individual cells were being passed back into visitTree, and when the visit occured, nothing was found, since they're not valid to visit.  The fix involves passing a parameter to AjaxBehaviorRenderer, which signals adding the client id to the rendered ajax call, instead of using the default, which is the individual cell.

Added tests to ensure no regressions.

r=rogerk

git-svn-id: https://svn.java.net/svn/mojarra~svn@7499 761efcf2-d34d-6b61-9145-99dcacc3edf1