Fix for bug 1219.
Change the default ajax behavior for all radio and checkbox components to be onclick instead of onchange, to allow for the odd behavior of IE.

Note that this means that this behavior will no longer  be "onchange" type behavior in other browsers either - if you select a radio button that's already selected, it will fire an event.  But the old behavior under IE was unacceptable.

I'll be sending an email to the EG about this change to notify them - however, this behavior is as close to the spec as we can currently make it, and thus should not be an issue.  No spec issue will be filed.
-------------

M CheckboxRenderer.java
M RadioRenderer.java
M SelectManyCheckboxListRenderer.java

Modify the renderers to call a new method in RenderKitUtils.

M RenderKitUtils.java

Add new method to create onclick action with a behaviorName of valueChanged

A SelectChangeBean.java
A browserSelectChange.xhtml 

Add new test page in the browser compat test suite, for eventual inclusion into QE testing.  Not wired into systest.



git-svn-id: https://svn.java.net/svn/mojarra~svn@7682 761efcf2-d34d-6b61-9145-99dcacc3edf1