Bug 1339 - MenuRenderer didn't correctly output options on Ajax requests
The fix for 1315 broke MenuRenderer, since it was using buffering for the options.  The fix was to use event buffering, instead of string buffering, so that we're still writing directly to the PartialResponseWriter.

RecordingResponseWriter - instead of buffing strings, this class buffers ResponseWriter events.  This lets us later replay the events into a different writer.

MenuRenderer - modify to use the RecordingResponseWriter.  

Add tests for regression.

git-svn-id: https://svn.java.net/svn/mojarra~svn@7921 761efcf2-d34d-6b61-9145-99dcacc3edf1