Fix for issue 1443.
Encode resources targeted in the 'head' within encodeBegin().

Per the issue description:

"CSS selectors placed in <h:head /> cannot override styles in implicitly added resources. Here's an 
example;

<h:head>
   <style>
      .selector1 {
       }
   </style>
</h:head>

If a component is adding a css using @ResourceDependency that has .selector1, it will take the 
precedence instead of the one in style tag. HeadRenderer uses encodeEnd to render viewroot resources,  
if it is done at encodeBegin, it'll become much easier to override these styles."

git-svn-id: https://svn.java.net/svn/mojarra~svn@8225 761efcf2-d34d-6b61-9145-99dcacc3edf1