Issue: 1512
No-one gave r= on this, but Neil Griffin did review the approach and
said it would not break portlets.

SECTION: SPEC Considerations

First, let's assume I'll file a spec issue, targeted at 2.0 Rev a, that
loosens the the following wording on the javadoc of
FacesServlet.service().

The very last sentance in the javadoc for FacesServlet.service() is:

"In a finally block, FacesContext.release() must be called."

I want to change this to:

"The implementation must make it so FacesContext.release() is called
within a finally block as late as possible in the processing for the JSF
related portion of this request".

SECTION: Impl considerations

I plan to affect this change by removing the call to
FacesContext.release() from FacesServlet.service() and moving it to a
finally block in Mojarra's implementation of
javax.servlet.ServletRequestListener.requestDestroyed().

Ryan Lubke mentioned has reason to believe that this method is not
called in a JSR-168 portlet environment.  Rather, he asserts we can only
rely on ServletContextListener in JSR-168.  

Neil, can you please share your insights about whether or not it is ok
to make this implementation change with respect to JSR-168 portlets?

Thanks,

Ed


SECTION: Modified Files
----------------------------
M       jsf-api/src/main/java/javax/faces/webapp/FacesServlet.java
M       jsf-ri/src/com/sun/faces/application/WebappLifecycleListener.java



git-svn-id: https://svn.java.net/svn/mojarra~svn@8293 761efcf2-d34d-6b61-9145-99dcacc3edf1