Performance improvemnt for cache impl https://javaserverfaces.dev.java.net/issues/show_bug.cgi?id=1613
Author: Blake Sullivan
r=edburns



SECTION: Modified Files
----------------------------
M       jsf-ri/src/main/java/com/sun/faces/util/Cache.java

Switch implementation to use ConcurrentHashMap.  I suspect that many of 
the current uses of cache don't care if the identify of the values in 
the cache are maintained.  To support those cases, an additional flag 
could be passed when creating the cache and put() could be used instead 
of the much slower putIfAbsent(), but since the caches see very few 
puts(), it probably doesn't matter.


git-svn-id: https://svn.java.net/svn/mojarra~svn@8527 761efcf2-d34d-6b61-9145-99dcacc3edf1