http://java.net/jira/browse/JAVASERVERFACES-2400
SECTION: Modified Files
----------------------------
M       quality/pom.xml

- Refer to the excludes

A       quality/findbugs-exclude.xml

- Exclude 

+  <Match>
+    <Field type="java.util.concurrent.ConcurrentHashMap" />
+    <Bug pattern="SE_BAD_FIELD" />
+  </Match>

This is safe to exclude because I found that when I removed the generics
information from the fields in question, I observed the SE_BAD_FIELD
message went away.  This observation indicates a bug in findbugs but we
are beholden to whatever version of findbugs the maven-findbugs-plugin
provides.  Even if we were to upgrade our version of the
maven-findbugs-plugin, it would still just use findbugs 2.0.  Therefore,
the best strategy is to exclude this error.


git-svn-id: https://svn.java.net/svn/mojarra~svn@10171 761efcf2-d34d-6b61-9145-99dcacc3edf1